user-service/requirements.txt
fastapi
uvicorn[standard]
aiokafka
sqlalchemy
psycopg2-binary
alembic 
pydantic
kafka-python
PyJWT
python-multipart